+++ Domain/Helpers/Formatter.cs
@@ -2,7 +2,7 @@ namespace SimplifiedSlotMachineV1.Domain.Helpers;
 
 public static class Formatter
 {
-    public static string Format(double item)
+    public static string Format(decimal item)
     {
         return item.ToString("0.0", System.Globalization.CultureInfo.InvariantCulture);
     }
